backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
httpx==0.27.2
